profile page for retailer's app
This repo contains updating a profile page. The profile can be updated as per the user requirements. The registering of the user be uploaded into the mongodb database. 

Here I have used the frontend code for checking the requirements instead of passport or any other.. You can find the frontend code in views section